Added my evaluation by NI in 2019
I took this test in 2019 in Brazil, I was 26 years old and had 7 years of experience using LabVIEW daily. I had already taken the Core 1, 2 and 3 courses - they are not mandatory, but they formalize knowledge. The test is long and required good time control to cover all the points.
Recertification can be done by recycling knowledge on the NI website, which is what I did, so I am not aware of any subsequent changes.